refactor(AndroidJob): Factorize into EventUploader impl
